+++ modules/tools/include/ifm3d/tools/cmdline_app.h
@@ -49,8 +49,8 @@ namespace ifm3d
     virtual bool CheckCompatibility();
 
     /**
-    * Method to execute ifm3d commands
-    */
+     * Method to execute ifm3d commands
+     */
     static int Execute(size_t argc, const char** argv);
 
   protected: